refactor: use createSlice for network reducer
